split Dracula files
