+++ config/splat.us.stst0.yaml
@@ -49,6 +49,8 @@ segments:
       - [0x27CAC, rodata]
       - [0x27CC4, rodata]
       - [0x27D64, c]
+      - [0x2C564, c] # EntityDracula
+      - [0x30030, c]
       - [0x36358, c]
       - [0x3EB00, sbss]
   - [0x425C4]
