+++ src/st/st0/30030.c
@@ -1,0 +1,854 @@
+#include "st0.h"
+
+void SetGameState(GameState gameState) {
+    g_GameState = gameState;
+    g_GameStep = 0;
+    g_backbufferX = 0;
+    g_backbufferY = 0;
+}
+
+void func_801B0058(void) {
+    s32 index = 0;
+    s32 i;
+    s32 j;
+
+    for (i = 0xF0; i < 0x100; i++) {
+        for (j = 0x200; j < 0x300; j += 0x10) {
+            D_8003C104[index++] = GetClut(j, i);
+        }
+    }
+
+    for (i = 0xF0; i < 0x100; i++) {
+        for (j = 0; j < 0x100; j += 0x10) {
+            D_8003C104[index++] = GetClut(j, i);
+        }
+    }
+
+    for (i = 0xF0; i < 0x100; i++) {
+        for (j = 0x100; j < 0x200; j += 0x10) {
+            D_8003C104[index++] = GetClut(j, i);
+        }
+    }
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B0180);
+
+void func_801B01C0(void) {
+    g_GpuBuffers[0].draw.r0 = 0;
+    g_GpuBuffers[0].draw.g0 = 0;
+    g_GpuBuffers[0].draw.b0 = 0;
+    g_GpuBuffers[1].draw.r0 = 0;
+    g_GpuBuffers[1].draw.g0 = 0;
+    g_GpuBuffers[1].draw.b0 = 0;
+}
+
+void func_801B01F8(s32 arg0) {
+    g_GpuBuffers[0].draw.clip.y = 0x0014;
+    g_GpuBuffers[0].draw.clip.h = 0x00CF;
+#ifndef NON_MATCHING
+    g_GpuBuffers_1_buf_draw_clip_y = arg0 == 0 ? 0x0014 : 0x0114;
+#else
+    g_GpuBuffers[1].draw.clip.y = 0x0014;
+#endif
+    g_GpuBuffers[1].draw.clip.h = 0x00CF;
+    g_GpuBuffers[1].draw.isbg = 1;
+    g_GpuBuffers[0].draw.isbg = 1;
+    func_801B01C0();
+    g_GpuBuffers[1].disp.isrgb24 = 0;
+    g_GpuBuffers[0].disp.isrgb24 = 0;
+}
+
+// Set stage display buffer
+void func_801B0280(void) {
+    SetDefDrawEnv(&g_GpuBuffers[0].draw, 0, 0, DISP_STAGE_W, DISP_STAGE_H);
+    SetDefDrawEnv(&g_GpuBuffers[1].draw, DISP_STAGE_W, 0, DISP_STAGE_W,
+                  DISP_STAGE_H);
+    SetDefDispEnv(&g_GpuBuffers[0].disp, DISP_STAGE_W, 0, DISP_STAGE_W,
+                  DISP_STAGE_H);
+    SetDefDispEnv(&g_GpuBuffers[1].disp, 0, 0, DISP_STAGE_W, DISP_STAGE_H);
+    func_801B01F8(0);
+}
+
+void func_801B0324(void) {
+    SetDefDrawEnv(&g_GpuBuffers[0].draw, 0, 0, DISP_UNK2_W, DISP_UNK2_H);
+    SetDefDrawEnv(&g_GpuBuffers[1].draw, 0, 256, DISP_UNK2_W, DISP_UNK2_H);
+    SetDefDispEnv(&g_GpuBuffers[0].disp, 0, 256, DISP_UNK2_W, DISP_UNK2_H);
+    SetDefDispEnv(&g_GpuBuffers[1].disp, 0, 0, DISP_UNK2_W, DISP_UNK2_H);
+    g_GpuBuffers[1].draw.clip.y = DISP_UNK2_W / 2;
+    g_GpuBuffers[1].draw.clip.h = DISP_UNK2_H;
+    g_GpuBuffers[0].draw.clip.h = DISP_UNK2_H;
+    g_GpuBuffers[0].draw.clip.y = 0;
+    g_GpuBuffers[1].draw.isbg = 1;
+    g_GpuBuffers[0].draw.isbg = 1;
+    func_801B01C0();
+    g_GpuBuffers[1].disp.isrgb24 = 0;
+    g_GpuBuffers[0].disp.isrgb24 = 0;
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B0414);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B0464);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B101C);
+
+void func_801B1198(s16 arg0) {
+    RECT rect;
+    rect.x = 0x200;
+    rect.y = arg0 * 16;
+    rect.w = 0x80;
+    rect.h = 0x10;
+
+    ClearImage(&rect, 0, 0, 0);
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B11E8);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B1298);
+
+s32 Random(void) {
+    g_randomNext = (g_randomNext * 0x01010101) + 1;
+    return g_randomNext >> 0x18;
+}
+
+void Update(void) {
+    s16 i;
+    Entity* entity;
+    s32* unk;
+
+    for (i = 0; i < 0x20; i++) {
+        if (D_801C2584[i]) {
+            D_801C2584[i]--;
+        }
+    }
+
+    unk = &D_80097410;
+    if (*unk) {
+        if (!--*unk) {
+            g_api.FreePrimitives(D_80097414);
+        }
+    }
+
+    for (entity = &g_Entities[STAGE_ENTITY_START];
+         entity < &g_Entities[TOTAL_ENTITY_COUNT]; entity++) {
+        if (!entity->pfnUpdate)
+            continue;
+
+        if (entity->step) {
+            s32 unk34 = entity->flags;
+            if (unk34 & FLAG_DESTROY_IF_OUT_OF_CAMERA) {
+                s16 posX = i = entity->posX.i.hi;
+                s16 posY = entity->posY.i.hi;
+                if (unk34 & FLAG_DESTROY_IF_BARELY_OUT_OF_CAMERA) {
+                    if ((u16)(posX + 64) > 384 || (u16)(posY + 64) > 352) {
+                        DestroyEntity(entity);
+                        continue;
+                    }
+                } else {
+                    if ((u16)(posX + 128) > 512 || (u16)(posY + 128) > 480) {
+                        DestroyEntity(entity);
+                        continue;
+                    }
+                }
+            }
+
+            if ((unk34 & 0x02000000)) {
+                s16 posY = entity->posY.i.hi + g_Camera.posY.i.hi;
+                s16 test = (g_CurrentRoom.vSize * 256) + 128;
+                if (posY > test) {
+                    DestroyEntity(entity);
+                    continue;
+                }
+            }
+
+            if (unk34 & 0xF) {
+                entity->palette =
+                    D_801815EC[(entity->unk49 << 1) | (unk34 & 1)];
+                entity->flags--;
+                if ((entity->flags & 0xF) == 0) {
+                    entity->palette = entity->unk6A;
+                    entity->unk6A = 0;
+                }
+            }
+
+            if (!(unk34 & 0x20000000) || (unk34 & 0x10000000) ||
+                ((u16)(entity->posX.i.hi + 64) <= 384) &&
+                    ((u16)(entity->posY.i.hi + 64) <= 352)) {
+                if (!entity->unk58 || (entity->unk58--, unk34 & 0x100000)) {
+                    if (!D_800973FC || unk34 & 0x2100 ||
+                        (unk34 & 0x200 && !(D_8003C8C4 & 3))) {
+                        g_CurrentEntity = entity;
+                        entity->pfnUpdate(entity);
+                        entity->unk44 = 0;
+                        entity->unk48 = 0;
+                    }
+                }
+            }
+        } else {
+            g_CurrentEntity = entity;
+            entity->pfnUpdate(entity);
+            entity->unk44 = 0;
+            entity->unk48 = 0;
+        }
+    }
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B1B98);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B1CA0);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", EntityNumericDamage);
+
+void CreateEntityFromLayout(Entity* entity, LayoutObject* initDesc) {
+    DestroyEntity(entity);
+    entity->objectId = initDesc->objectId & 0x3FF;
+    entity->pfnUpdate = PfnEntityUpdates[entity->objectId];
+    entity->posX.i.hi = initDesc->posX - g_Camera.posX.i.hi;
+    entity->posY.i.hi = initDesc->posY - g_Camera.posY.i.hi;
+    entity->subId = initDesc->subId;
+    entity->objectRoomIndex = initDesc->objectRoomIndex >> 8;
+    entity->unk68 = (initDesc->objectId >> 0xA) & 7;
+}
+
+void CreateEntityWhenInVerticalRange(LayoutObject* layoutObj) {
+    s16 yClose;
+    s16 yFar;
+    s16 posY;
+    Entity* entity;
+
+    posY = g_Camera.posY.i.hi;
+    yClose = posY - 0x40;
+    yFar = posY + 0x120;
+    if (yClose < 0) {
+        yClose = 0;
+    }
+
+    posY = layoutObj->posY;
+    if (posY < yClose) {
+        return;
+    }
+
+    if (yFar < posY) {
+        return;
+    }
+
+    switch (layoutObj->objectId & 0xE000) {
+    case 0x0:
+        entity =
+            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
+        if (entity->objectId == 0) {
+            CreateEntityFromLayout(entity, layoutObj);
+        }
+        break;
+    case 0x8000:
+        break;
+    case 0xA000:
+        entity =
+            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
+        CreateEntityFromLayout(entity, layoutObj);
+        break;
+    }
+}
+
+void CreateEntityWhenInHorizontalRange(LayoutObject* layoutObj) {
+    s16 xClose;
+    s16 xFar;
+    s16 posX;
+    Entity* entity;
+
+    posX = g_Camera.posX.i.hi;
+    xClose = posX - 0x40;
+    xFar = posX + 0x140;
+    if (xClose < 0) {
+        xClose = 0;
+    }
+
+    posX = layoutObj->posX;
+    if (posX < xClose) {
+        return;
+    }
+
+    if (xFar < posX) {
+        return;
+    }
+
+    switch (layoutObj->objectId & 0xE000) {
+    case 0x0:
+        entity =
+            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
+        if (entity->objectId == 0) {
+            CreateEntityFromLayout(entity, layoutObj);
+        }
+        break;
+    case 0x8000:
+        break;
+    case 0xA000:
+        entity =
+            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
+        CreateEntityFromLayout(entity, layoutObj);
+        break;
+    }
+}
+
+extern LayoutObject* D_801C00A0;
+void func_801B33D4(s16 arg0) {
+    while (true) {
+        if (D_801C00A0->posX != 0xFFFE && D_801C00A0->posX >= arg0) {
+            break;
+        }
+        D_801C00A0++;
+    }
+}
+
+void func_801B3420(s16 arg0) {
+    while (true) {
+        if (D_801C00A0->posX != 0xFFFF &&
+            (arg0 >= D_801C00A0->posX || D_801C00A0->posX == 0xFFFE)) {
+            break;
+        }
+        D_801C00A0--;
+    }
+}
+
+void func_801B3478(s16);
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B3478);
+
+void func_801B3574(s16);
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B3574);
+
+extern LayoutObject* D_801C00A4;
+void func_801B3688(s16 arg0) {
+    while (true) {
+        if (D_801C00A4->posY != 0xFFFE && (s32)D_801C00A4->posY >= arg0) {
+            break;
+        }
+        D_801C00A4++;
+    }
+}
+
+void func_801B36D4(s16 arg0) {
+    while (true) {
+        if ((D_801C00A4->posY != 0xFFFF) &&
+            ((arg0 >= D_801C00A4->posY) || (D_801C00A4->posY == 0xFFFE))) {
+            break;
+        }
+        D_801C00A4--;
+    }
+}
+
+void func_801B372C(s16);
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B372C);
+
+void func_801B3828(s16);
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B3828);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", InitRoomEntities);
+
+void func_801B3AB4(void) {
+    Unkstruct8* currentRoomTileLayout = &g_CurrentRoomTileLayout;
+
+    if (D_80097908 != 0) {
+        s16 tmp = g_Camera.posX.i.hi;
+        if (D_80097908 > 0)
+            func_801B3478(tmp + 320);
+        else
+            func_801B3574(tmp - 64);
+    }
+
+    if (D_8009790C != 0) {
+        s16 tmp = currentRoomTileLayout->unkE;
+        if (D_8009790C > 0)
+            func_801B372C(currentRoomTileLayout->unkE + 288);
+        else
+            func_801B3828(tmp - 64);
+    }
+}
+
+void CreateEntityFromCurrentEntity(u16 objectId, Entity* entity) {
+    DestroyEntity(entity);
+    entity->objectId = objectId;
+    entity->pfnUpdate = PfnEntityUpdates[objectId];
+    entity->posX.i.hi = g_CurrentEntity->posX.i.hi;
+    entity->posY.i.hi = g_CurrentEntity->posY.i.hi;
+}
+
+void CreateEntityFromEntity(u16 objectId, Entity* source, Entity* entity) {
+    DestroyEntity(entity);
+    entity->objectId = objectId;
+    entity->pfnUpdate = PfnEntityUpdates[objectId];
+    entity->posX.i.hi = source->posX.i.hi;
+    entity->posY.i.hi = source->posY.i.hi;
+}
+
+s32 func_801B3C58(Entity* e) {
+    s16 diff;
+
+    diff = PLAYER.posX.i.hi - e->posX.i.hi;
+    diff = ABS(diff);
+
+    if (diff >= 17) {
+        diff = 0;
+    } else {
+        diff = PLAYER.posY.i.hi - e->posY.i.hi;
+        diff = ABS(diff);
+        diff = diff < 33;
+    }
+
+    return diff;
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", EntityRedDoor);
+
+void DestroyEntity(Entity* item) {
+    s32 i;
+    s32 length;
+    u32* ptr;
+
+    if (item->flags & FLAG_FREE_POLYGONS) {
+        g_api.FreePrimitives(item->firstPolygonIndex);
+    }
+
+    ptr = (u32*)item;
+    length = sizeof(Entity) / sizeof(s32);
+    for (i = 0; i < length; i++)
+        *ptr++ = 0;
+}
+
+void DestroyEntityFromIndex(s16 index) {
+    Entity* entity = &g_Entities[index];
+
+    while (entity < &D_8007EF1C) {
+        DestroyEntity(entity);
+        entity++;
+    }
+}
+
+void PreventEntityFromRespawning(Entity* entity) {
+    if (entity->objectRoomIndex) {
+        u32 value = (entity->objectRoomIndex - 1);
+        u16 index = value / 32;
+        u16 bit = value % 32;
+        g_entityDestroyed[index] |= 1 << bit;
+    }
+}
+
+#include "st/AnimateEntity.h"
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B4AF0);
+
+/*
+ * Returns the absolute distance from g_CurrentEntity to player in the X Axis
+ */
+s16 func_801B4C08(void) {
+    s16 xDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
+
+    if (xDistance < 0) {
+        xDistance = -xDistance;
+    }
+    return xDistance;
+}
+
+/*
+ * Returns the absolute distance from g_CurrentEntity to player in the Y Axis
+ */
+s32 func_801B4C44(void) {
+    s32 yDistance = g_CurrentEntity->posY.i.hi - PLAYER.posY.i.hi;
+
+    if (yDistance < 0) {
+        yDistance = -yDistance;
+    }
+    return yDistance;
+}
+
+s16 func_801B4C78(void) {
+    s16 var_a0 = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
+
+    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
+        var_a0 |= 2;
+    }
+    return var_a0;
+}
+
+void MoveEntity(void) {
+    g_CurrentEntity->posX.val += g_CurrentEntity->accelerationX;
+    g_CurrentEntity->posY.val += g_CurrentEntity->accelerationY;
+}
+
+void FallEntity(void) {
+    if (g_CurrentEntity->accelerationY < FALL_TERMINAL_VELOCITY) {
+        g_CurrentEntity->accelerationY += FALL_GRAVITY;
+    }
+}
+
+u8 func_801B4D18(void) {
+    u8 unkState;
+    Entity* entity;
+
+    MoveEntity();
+    FallEntity();
+
+    entity = g_CurrentEntity;
+
+    if (unkState & 1) {
+        entity->accelerationY = 0;
+    }
+
+    return unkState;
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B4D5C);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B4FD4);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B51E4);
+
+Entity* AllocEntity(Entity* start, Entity* end) {
+    Entity* current = start;
+    while (current < end) {
+        if (current->objectId == 0) {
+            DestroyEntity(current);
+            return current;
+        }
+
+        current++;
+    }
+    return NULL;
+}
+
+s32 func_801B542C(u8 arg0, s16 arg1) { return D_80181990[arg0] * arg1; }
+
+s16 func_801B5458(u8 arg0) { return D_80181990[arg0]; }
+
+void func_801B5474(s32 arg0, s16 arg1) {
+    g_CurrentEntity->accelerationX = func_801B542C(arg0, arg1);
+    g_CurrentEntity->accelerationY = func_801B542C(arg0 - 0x40, arg1);
+}
+
+u8 func_801B54E0(s16 arg0, s16 arg1) {
+    return ((ratan2(arg1, arg0) >> 4) + 0x40);
+}
+
+u8 func_801B5518(Entity* arg0, Entity* arg1) {
+    s16 a = arg1->posX.i.hi - arg0->posX.i.hi;
+    s16 b = arg1->posY.i.hi - arg0->posY.i.hi;
+    return func_801B54E0(a, b);
+}
+
+u8 func_801B5560(s32 arg0, s32 arg1) {
+    s16 a = (arg0 - (u16)g_CurrentEntity->posX.i.hi);
+    s16 b = (arg1 - (u16)g_CurrentEntity->posY.i.hi);
+    return func_801B54E0(a, b);
+}
+
+u8 func_801B55A8(u8 arg0, u8 arg1, u8 arg2) {
+    u8 var_v0;
+    s8 temp_a2 = arg2 - arg1;
+
+    if (temp_a2 < 0) {
+        var_v0 = -temp_a2;
+    } else {
+        var_v0 = temp_a2;
+    }
+
+    if (var_v0 > arg0) {
+        if (temp_a2 < 0) {
+            var_v0 = arg1 - arg0;
+        } else {
+            var_v0 = arg1 + arg0;
+        }
+
+        return var_v0;
+    }
+
+    return arg2;
+}
+
+void func_801B5600(u16 slope, s16 speed) {
+    Entity* entity;
+    s32 moveX;
+    s32 moveY;
+
+    moveX = rcos(slope) * speed;
+    entity = g_CurrentEntity;
+
+    if (moveX < 0) {
+        moveX += 15;
+    }
+
+    entity->accelerationX = moveX >> 4;
+
+    moveY = rsin(slope) * speed;
+    entity = g_CurrentEntity;
+
+    if (moveY < 0) {
+        moveY += 15;
+    }
+
+    entity->accelerationY = moveY >> 4;
+}
+
+u16 func_801B568C(s16 x, s16 y) { return ratan2(y, x); }
+
+u16 func_801B56BC(Entity* a, Entity* b) {
+    s32 diffX = b->posX.i.hi - a->posX.i.hi;
+    s32 diffY = b->posY.i.hi - a->posY.i.hi;
+    return ratan2(diffY, diffX);
+}
+
+u16 func_801B56F4(s32 x, s32 y) {
+    s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
+    s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
+    return ratan2(diffY, diffX);
+}
+
+// minor reg swap
+#ifndef NON_MATCHING
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B573C);
+#else
+u16 func_801B573C(u16 arg0, s16 arg1, s16 arg2) {
+    u16 temp_a2 = arg2 - arg1;
+    u16 var_v0 = arg1;
+    u16 var_v0_2;
+
+    if (temp_a2 & 0x800) {
+        var_v0_2 = temp_a2 & 0x7FF;
+    } else {
+        var_v0_2 = temp_a2;
+    }
+
+    if (var_v0_2 > arg0) {
+        if (temp_a2 & 0x800) {
+            var_v0 = arg1 - arg0;
+        } else {
+            var_v0 = arg1 + arg0;
+        }
+
+        return var_v0;
+    }
+
+    return arg2;
+}
+#endif
+
+void func_801B5794(u8 state) {
+    g_CurrentEntity->step = state;
+    g_CurrentEntity->step_s = 0;
+    g_CurrentEntity->animFrameIdx = 0;
+    g_CurrentEntity->animFrameDuration = 0;
+}
+
+void func_801B57B4(u8 arg0) {
+    g_CurrentEntity->step_s = arg0;
+    g_CurrentEntity->animFrameIdx = 0;
+    g_CurrentEntity->animFrameDuration = 0;
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B57D0);
+
+void InitializeEntity(u16 arg0[]) {
+    u16 enemyId;
+    EnemyDef* enemyDef;
+
+    g_CurrentEntity->animSet = *arg0++;
+    g_CurrentEntity->animCurFrame = *arg0++;
+    g_CurrentEntity->unk5A = *arg0++;
+    g_CurrentEntity->palette = *arg0++;
+
+    enemyId = *arg0++;
+    g_CurrentEntity->enemyId = enemyId;
+    enemyDef = &g_api.enemyDefs[enemyId];
+    g_CurrentEntity->hitPoints = enemyDef->hitPoints;
+    g_CurrentEntity->attack = enemyDef->attack;
+    g_CurrentEntity->attackElement = enemyDef->attackElement;
+    g_CurrentEntity->unk3C = enemyDef->unkC;
+    g_CurrentEntity->hitboxWidth = enemyDef->hitboxWidth;
+    g_CurrentEntity->hitboxHeight = enemyDef->hitboxHeight;
+    g_CurrentEntity->flags = enemyDef->unk24;
+    g_CurrentEntity->unk10 = 0;
+    g_CurrentEntity->unk12 = 0;
+    g_CurrentEntity->step_s = 0;
+    g_CurrentEntity->step++;
+    if (g_CurrentEntity->zPriority == 0) {
+        g_CurrentEntity->zPriority = g_zEntityCenter.S16.unk0 - 0xC;
+    }
+}
+
+void EntityDummy(Entity* arg0) {
+    if (arg0->step == 0) {
+        arg0->step++;
+    }
+}
+
+s32 func_801B5970(u16* hitSensors, s16 sensorCount) {
+    Collider collider;
+    s16 i;
+    s32 accelerationX;
+    u16 temp_a1;
+    s16 x;
+    s16 y;
+
+    accelerationX = g_CurrentEntity->accelerationX;
+    if (accelerationX != 0) {
+        x = g_CurrentEntity->posX.i.hi;
+        y = g_CurrentEntity->posY.i.hi;
+        for (i = 0; i < sensorCount; i++) {
+            if (accelerationX < 0) {
+                s16 newX = x + *hitSensors++;
+                x = newX;
+            } else {
+                s16 newX = x - *hitSensors++;
+                x = newX;
+            }
+
+            y += *hitSensors++;
+            g_api.CheckCollision(x, y, &collider, 0);
+            if (collider.unk0 & 2 &&
+                ((!(collider.unk0 & 0x8000)) || (i != 0))) {
+                return 2;
+            }
+        }
+        return 0;
+    }
+}
+
+void func_801B5A98(u16* hitSensors, s16 sensorCount) {
+    Collider collider;
+    s16 i;
+    s32 accelerationX;
+    s16 x;
+    s16 y;
+
+    accelerationX = g_CurrentEntity->accelerationX;
+    if (accelerationX == 0)
+        return;
+    x = g_CurrentEntity->posX.i.hi;
+    y = g_CurrentEntity->posY.i.hi;
+    for (i = 0; i < sensorCount; i++) {
+        if (accelerationX < 0) {
+            x = x + *hitSensors++;
+        } else {
+            x = x - *hitSensors++;
+        }
+
+        y += *hitSensors++;
+        g_api.CheckCollision(x, y, &collider, 0);
+        if (collider.unk0 & 2 && (!(collider.unk0 & 0x8000) || i != 0)) {
+            if (accelerationX < 0) {
+                g_CurrentEntity->posX.i.hi += collider.unk1C;
+            } else {
+                g_CurrentEntity->posX.i.hi += collider.unk14;
+            }
+            return;
+        }
+    }
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", func_801B5BF0);
+
+void ReplaceBreakableWithItemDrop(Entity* self) {
+    u16 subId;
+
+    PreventEntityFromRespawning(self);
+
+#if STAGE != STAGE_ST0
+    if (!(g_Status.relics[10] & 2)) {
+        DestroyEntity(self);
+        return;
+    }
+#endif
+
+    subId = self->subId &= 0xFFF;
+
+    if (subId < 0x80) {
+        self->objectId = E_PRIZE_DROP;
+        self->pfnUpdate = (PfnEntityUpdate)EntityPrizeDrop;
+        self->animFrameDuration = 0;
+        self->animFrameIdx = 0;
+    } else {
+        subId -= 0x80;
+        self->objectId = E_EQUIP_ITEM_DROP;
+        self->pfnUpdate = (PfnEntityUpdate)EntityEquipItemDrop;
+    }
+
+    self->subId = subId;
+    self->unk6D = 0x10;
+    self->step = 0;
+}
+
+void func_801B5EC8(void) {
+    s32 temp_v1;
+    Entity* entity;
+
+    entity = g_CurrentEntity;
+    if (entity->accelerationY >= 0) {
+        temp_v1 =
+            entity->ext.generic.unk88.S16.unk0 + entity->ext.generic.unk84.unk;
+        entity->ext.generic.unk84.unk = temp_v1;
+        entity->accelerationX = temp_v1;
+        if (temp_v1 == 0x10000 || temp_v1 == -0x10000) {
+            entity->ext.generic.unk88.S16.unk0 =
+                -entity->ext.generic.unk88.S16.unk0;
+        }
+        entity = g_CurrentEntity;
+    }
+
+    if (entity->accelerationY < 0x00004000) {
+        entity->accelerationY += 0x2000;
+    }
+}
+
+void func_801B5F4C(u16 arg0) {
+    Collider res;
+
+    if (g_CurrentEntity->accelerationX < 0) {
+        g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
+                             g_CurrentEntity->posY.i.hi - 7, &res, 0);
+        if (res.unk0 & 5) {
+            g_CurrentEntity->accelerationY = 0;
+        }
+    }
+
+    g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
+                         g_CurrentEntity->posY.i.hi + 7, &res, 0);
+
+    if (arg0) {
+        if (!(res.unk0 & 5)) {
+            MoveEntity();
+            FallEntity();
+            return;
+        }
+
+        g_CurrentEntity->accelerationX = 0;
+        g_CurrentEntity->accelerationY = 0;
+
+        if (res.unk0 & 4) {
+            g_CurrentEntity->posY.val += 0x2000;
+            return;
+        }
+
+        g_CurrentEntity->posY.i.hi += res.unk18;
+        return;
+    }
+
+    if (!(res.unk0 & 5)) {
+        MoveEntity();
+        func_801B5EC8();
+    }
+}
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", CollectHeart);
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", CollectGold);
+
+void func_801B61D4(void) {}
+
+void func_801B61DC(void) { DestroyEntity(g_CurrentEntity); }
+
+INCLUDE_ASM("asm/us/st/st0/nonmatchings/30030", CollectSubweapon);
+
+void CollectDummy(void) { DestroyEntity(g_CurrentEntity); }
+
+Entity* func_801B633C(void) {
+    g_CurrentEntity->step = 3;
+    g_CurrentEntity->subId = 4;
+    return g_CurrentEntity;
+}
